Improve check coverage about idle threads.
Idle threads are not allowed to acquire any lock but spinlocks.
Deny any attempt to do so by panicing at the locking operation
when INVARIANTS is on. Then, remove the check on blocking on a
turnstile.
The check in sleepqueues is left because they are not allowed to use
tsleep() either which could happen still.

Reviewed by:	bde, jhb, kib
MFC after:	1 week

(cherry picked from commit a634dfc320608f9955ab0350a7f78ee73064b909)

Conflicts:
	sys/kern/kern_mutex.c
	sys/kern/kern_rwlock.c
	sys/kern/kern_sx.c