MFC r255047 (with a manual tweak to match the ext_free in stable/9):
Add a routine for attaching an mbuf to a buffer with an external
refcount.  This one is willing to work with buffers that may already be
referenced.  MEXTADD/m_extadd are suitable only for the first attachment
to a cluster -- they initialize the refcount to 1.

(cherry picked from commit 77304f3b509fbfe1c56c8ff49979f6a89ca36b46)